Added feature off sending, accepting friend requests
