+++ project_garage/Controllers/ProfileController.cs
@@ -52,123 +52,123 @@ namespace project_garage.Controllers
             return View(viewModel);
         }
 
-        [HttpPost]
-        [Route("Friends/Add")]
-        public async Task<IActionResult> AddFriend(string friendId)
-        {
-            var userId = User.GetUserId();
+        //[HttpPost]
+        //[Route("Friends/Add")]
+        //public async Task<IActionResult> AddFriend(string friendId)
+        //{
+        //    var userId = User.GetUserId();
 
-            try
-            {
-                await _friendService.SendFriendRequestAsync(friendId, userId);
-                return Ok(new { message = "Friend request set successfully." });
-            }
-            catch (InvalidOperationException ex)
-            {
-                return BadRequest(new { error = ex.Message });
-            }
-            catch (Exception)
-            {
-                return StatusCode(500, new { error = "An error occurred while sending the friend request." });
-            }
+        //    try
+        //    {
+        //        await _friendService.SendFriendRequestAsync(friendId, userId);
+        //        return Ok(new { message = "Friend request set successfully." });
+        //    }
+        //    catch (InvalidOperationException ex)
+        //    {
+        //        return BadRequest(new { error = ex.Message });
+        //    }
+        //    catch (Exception)
+        //    {
+        //        return StatusCode(500, new { error = "An error occurred while sending the friend request." });
+        //    }
 
-        }
+        //}
 
-        [HttpPost]
-        [Route("Profile/Accept")]
-        public async Task<IActionResult> AcceptFriend(string friendId)
-        {
-            Console.WriteLine($"{friendId}");
+        //[HttpPost]
+        //[Route("Profile/Accept")]
+        //public async Task<IActionResult> AcceptFriend(string friendId)
+        //{
+        //    Console.WriteLine($"{friendId}");
 
-            var userId = User.GetUserId();
-            try
-            {
+        //    var userId = User.GetUserId();
+        //    try
+        //    {
 
-                await _friendService.AcceptRequestAsync(userId, friendId);
-            }
-            catch (InvalidOperationException ex)
-            {
-                return BadRequest(new { error = ex.Message });
-            }
-            catch (Exception)
-            {
-                return StatusCode(500, new { error = "An error occured while accepting request" });
-            }
+        //        await _friendService.AcceptRequestAsync(userId, friendId);
+        //    }
+        //    catch (InvalidOperationException ex)
+        //    {
+        //        return BadRequest(new { error = ex.Message });
+        //    }
+        //    catch (Exception)
+        //    {
+        //        return StatusCode(500, new { error = "An error occured while accepting request" });
+        //    }
 
-            return View();
-        }
+        //    return View();
+        //}
 
-        [HttpPost]
-        [Route("Profile/Reject")]
-        public async Task<IActionResult> RejectFriend(string friendId)
-        {
-            var userId = User.GetUserId();
-            try
-            {
-                await _friendService.RejectOrDeleteAsync(userId, friendId);
-            }
-            catch (InvalidOperationException ex)
-            {
-                return BadRequest(new { error = ex.Message });
-            }
-            catch (Exception)
-            { 
-                return StatusCode(500, new { error = "An error occured while rejecting" }); 
-            }
-            return View();
-        }
+        //[HttpPost]
+        //[Route("Profile/Reject")]
+        //public async Task<IActionResult> RejectFriend(string friendId)
+        //{
+        //    var userId = User.GetUserId();
+        //    try
+        //    {
+        //        await _friendService.RejectOrDeleteAsync(userId, friendId);
+        //    }
+        //    catch (InvalidOperationException ex)
+        //    {
+        //        return BadRequest(new { error = ex.Message });
+        //    }
+        //    catch (Exception)
+        //    { 
+        //        return StatusCode(500, new { error = "An error occured while rejecting" }); 
+        //    }
+        //    return View();
+        //}
 
-        [HttpGet]
-        [Route("Profile/GetAllRequests")]
-        public async Task<IActionResult> GetAllRequests()
-        {
-            var userId = User.GetUserId();
-            Console.WriteLine("-------------------------");
-            try
-            {
-                var requests = await _friendService.GetByUserIdAsync(userId);
+        //[HttpGet]
+        //[Route("Profile/GetAllRequests")]
+        //public async Task<IActionResult> GetAllRequests()
+        //{
+        //    var userId = User.GetUserId();
+        //    Console.WriteLine("-------------------------");
+        //    try
+        //    {
+        //        var requests = await _friendService.GetByUserIdAsync(userId);
 
-                var viewModel = new List<FriendRequestViewModel>();
+        //        var viewModel = new List<FriendRequestViewModel>();
 
-                foreach (var request in requests)
-                {
-                    if (!request.IsAccepted)
-                    {
-                        // Отримуємо дані друга через FriendId
-                        var friend = await _userService.GetByIdAsync(request.FriendId);
+        //        foreach (var request in requests)
+        //        {
+        //            if (!request.IsAccepted)
+        //            {
+        //                // Отримуємо дані друга через FriendId
+        //                var friend = await _userService.GetByIdAsync(request.FriendId);
 
-                        if (friend != null)
-                        {
-                            viewModel.Add(new FriendRequestViewModel
-                            {
-                                RequestId = request.Id,
-                                SenderId = friend.Id, // ID друга (відправника заявки)
-                                SenderName = friend.UserName, // Ім'я друга
-                                SenderDescription = friend.Description // Опис друга
-                            });
-                        }
-                    }
+        //                if (friend != null)
+        //                {
+        //                    viewModel.Add(new FriendRequestViewModel
+        //                    {
+        //                        RequestId = request.Id,
+        //                        SenderId = friend.Id, // ID друга (відправника заявки)
+        //                        SenderName = friend.UserName, // Ім'я друга
+        //                        SenderDescription = friend.Description // Опис друга
+        //                    });
+        //                }
+        //            }
 
-                    return View(viewModel); // Передаємо список у View
-                }
+        //            return View(viewModel); // Передаємо список у View
+        //        }
 
-            }
-            catch (InvalidOperationException ex)
-            {
-                return BadRequest(new { error = ex.Message });
-            }
-            catch (Exception)
-            {
-                return StatusCode(500, new { error = "An error occured while rejecting" });
-            }
-            return View();
-        }
+        //    }
+        //    catch (InvalidOperationException ex)
+        //    {
+        //        return BadRequest(new { error = ex.Message });
+        //    }
+        //    catch (Exception)
+        //    {
+        //        return StatusCode(500, new { error = "An error occured while rejecting" });
+        //    }
+        //    return View();
+        //}
 
-        [HttpPost]
-        public IActionResult GetAllRequests(FriendRequestViewModel model)
-        {
-            return View(model);
-        }
+        //[HttpPost]
+        //public IActionResult GetAllRequests(FriendRequestViewModel model)
+        //{
+        //    return View(model);
+        //}
 
         [HttpGet]
         public IActionResult SearchUsers()
