+++ project_garage/Views/Friend/GetUnacceptedRequests.cshtml
@@ -12,13 +12,13 @@
                 <strong>@request.SenderName</strong> (@request.SenderDescription)
                 <div>
                     <!-- Форма для прийняття заявки -->
-                    <form method="post" action="/Profile/Accept">
+                    <form method="post" action="/Friend/Accept">
                         <input type="hidden" name="friendId" value="@request.SenderId" />
                         <button type="submit" class="btn btn-success">Accept</button>
                     </form>
 
                     <!-- Форма для відхилення заявки -->
-                    <form method="post" action="/Profile/Reject">
+                    <form method="post" action="/Friend/Reject">
                         <input type="hidden" name="friendId" value="@request.SenderId" />
                         <button type="submit" class="btn btn-danger">Reject</button>
                     </form>
